move case assignment code to bihar custom app
