+++ custom/bihar/exceptions.py
